Refactoring
Extracted MapWithListView component
